Add the beta apollo hooks package and test it out on the welcome page. So far so good!
