+++ client/package.json
@@ -29,70 +29,71 @@
     "clean-build": "rm -rf build"
   },
   "dependencies": {
-    "@blueprintjs/core": "^3.13.0",
-    "@blueprintjs/icons": "^3.6.0",
-    "@sentry/browser": "^4.6.3",
+    "@apollo/react-hooks": "0.1.0-beta.2",
+    "@blueprintjs/core": "^3.15.1",
+    "@blueprintjs/icons": "^3.8.0",
+    "@sentry/browser": "^4.6.6",
     "@tweenjs/tween.js": "^17.3.0",
-    "apollo-cache-inmemory": "^1.4.3",
-    "apollo-client": "^2.4.13",
-    "apollo-link": "^1.2.8",
+    "apollo-cache-inmemory": "^1.5.1",
+    "apollo-client": "^2.5.1",
+    "apollo-link": "^1.2.11",
     "apollo-link-context": "^1.0.17",
-    "apollo-link-error": "^1.1.7",
-    "apollo-link-http": "^1.5.11",
-    "apollo-link-ws": "^1.0.14",
-    "apollo-utilities": "^1.1.3",
+    "apollo-link-error": "^1.1.10",
+    "apollo-link-http": "^1.5.14",
+    "apollo-link-ws": "^1.0.17",
+    "apollo-utilities": "^1.2.1",
     "appdirectory": "^0.1.0",
     "bodymovin": "^4.13.0",
     "bootstrap": "^4.3.1",
     "centered-pan-zoom": "^2.0.0",
     "change-case": "^3.1.0",
-    "color": "^3.1.0",
+    "color": "^3.1.1",
     "escape-string-regexp": "^1.0.5",
-    "graphql": "^14.1.1",
+    "graphql": "^14.3.0",
     "graphql-tag": "^2.10.1",
     "graphql-tag.macro": "^2.1.0",
     "http-proxy-middleware": "^0.19.1",
     "keycode": "^2.2.0",
-    "luxon": "^1.11.3",
+    "luxon": "^1.13.2",
     "papaparse": "^4.6.3",
     "prando": "^5.1.0",
     "random-words": "^1.1.0",
-    "rc-menu": "^7.4.21",
-    "react": "^16.8.3",
+    "rc-menu": "^7.4.22",
+    "react": "^16.8.6",
     "react-addons-transition-group": "^15.6.2",
     "react-animated-number": "^0.4.4",
-    "react-apollo": "^2.4.1",
-    "react-color": "^2.17.0",
-    "react-dom": "^16.8.3",
-    "react-draggable": "^3.1.1",
+    "react-apollo": "^2.5.5",
+    "react-color": "^2.17.3",
+    "react-dom": "^16.8.6",
+    "react-draggable": "^3.3.0",
     "react-fontawesome": "^1.6.1",
-    "react-icons": "^3.4.0",
+    "react-icons": "^3.7.0",
     "react-intl": "^2.8.0",
     "react-konami": "^0.5.0",
-    "react-measure": "^2.2.4",
-    "react-media-player": "^0.7.6",
+    "react-measure": "^2.3.0",
+    "react-media-player": "^0.7.7",
     "react-mosaic-component": "^2.1.0",
     "react-qrcode-hook": "^1.1.0",
     "react-rangeslider": "^2.2.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "3.0.0",
-    "react-sortable-hoc": "^1.6.1",
-    "react-table": "^6.9.2",
+    "react-sortable-hoc": "^1.9.1",
+    "react-table": "^6.10.0",
     "react-typing": "^1.0.0",
-    "reactour": "^1.13.1",
+    "reactour": "^1.13.4",
     "reactstrap": "^7.1.0",
     "sine-waves": "^0.3.0",
-    "styled-components": "^4.1.3",
-    "subscriptions-transport-ws": "^0.9.15",
+    "styled-components": "^4.2.0",
+    "subscriptions-transport-ws": "^0.9.16",
     "svg-to-jsx": "^1.0.0",
-    "terminal-in-react": "^4.3.0",
+    "terminal-in-react": "^4.3.1",
     "three": "^0.101.1",
     "three-obj-loader": "^1.1.3",
     "tinycolor2": "^1.4.1",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "eslint-plugin-react-hooks": "^1.2.0",
+    "eslint-plugin-react-hooks": "^1.6.0",
     "jest-localstorage-mock": "^2.4.0",
     "node-sass-chokidar": "^1.3.4",
     "preval.macro": "^3.0.0",
