+++ client/src/App.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { ApolloProvider } from "react-apollo";
+import { ApolloProvider as ApolloHooksProvider } from "@apollo/react-hooks";
 import client from "./helpers/graphqlClient";
 import Routes from "./containers/routes";
 import ErrorBoundary from "./helpers/errorBoundary";
@@ -10,12 +11,14 @@ import "./fonts.scss";
 
 const ApolloApp = () => (
   <ApolloProvider client={client}>
-    <IntlProvider>
-      <ErrorBoundary>
-        <Routes />
-        <EasterEgg />
-      </ErrorBoundary>
-    </IntlProvider>
+    <ApolloHooksProvider client={client}>
+      <IntlProvider>
+        <ErrorBoundary>
+          <Routes />
+          <EasterEgg />
+        </ErrorBoundary>
+      </IntlProvider>
+    </ApolloHooksProvider>
   </ApolloProvider>
 );
