+++ client/src/containers/FlightDirector/Welcome/welcome.scss
