+++ client/src/containers/FlightDirector/index.js
@@ -3,7 +3,7 @@ export { default as FlightDirector } from "./FlightDirector";
 export { default as MissionConfig } from "./MissionConfig";
 export { default as SimulatorConfig } from "./SimulatorConfig";
 export { default as AssetConfig } from "./AssetConfig";
-export { default as Welcome } from "./Welcome";
+export { default as Welcome } from "./Welcome/Welcome";
 export { default as ClientsLobby } from "./ClientsLobby";
 export { default as SetConfig } from "./SetConfig";
 export { default as SoftwarePanels } from "./SoftwarePanels";
